Single out SimpleTimeIntegrator
... For easier analysis and editing
